+++ assets/js/src/ColorPicker.tsx
@@ -1,10 +1,14 @@
-import React, { KeyboardEventHandler, useCallback } from "react";
+import React, { KeyboardEventHandler, useCallback, useRef } from "react";
 import { HexAlphaColorPicker } from "react-colorful";
 import useEyeDropper from "use-eye-dropper";
 import { BiSolidEyedropper } from "react-icons/bi";
+import { useDrag } from "@use-gesture/react";
+import { useSpring, animated, config } from "@react-spring/web";
+import { Cross2Icon } from "@radix-ui/react-icons";
 import { useAppDispatch, useAppSelector } from "./hooks";
 import {
   selectBrushColor,
+  selectIsColorPickerPosition,
   selectMaskColor,
   setBrushColor,
   setMaskColor,
@@ -18,7 +22,39 @@ const rgbaToHex = (rgba) => {
   return `#${((1 << 24) + (r << 16) + (g << 8) + b).toString(16).slice(1)}`;
 };
 
-const ColorPicker = () => {
+export type ColorPickerProps = {
+  onColorChange?: (color: string) => void;
+  onClose?: () => void;
+  position: { x: number; y: number };
+  isVisible?: boolean;
+};
+
+const ColorPicker = ({
+  position,
+  // isVisible,
+  onColorChange,
+  onClose,
+}: ColorPickerProps) => {
+  const colorPickerPosition = useAppSelector(selectIsColorPickerPosition);
+  const containerRef = useRef<HTMLDivElement>(null);
+  const posX = position.x ?? colorPickerPosition?.x;
+  const posY = position.y ?? colorPickerPosition?.y;
+  const [style, api] = useSpring(() => ({
+    x: 0,
+    y: 0,
+    left: posX,
+    top: posY,
+    config: config.stiff,
+    // opacity: isVisible ? 1 : 0,
+  }));
+  const bind = useDrag(({ offset: [x, y], target, cancel }) => {
+    if (containerRef.current === target) {
+      api.start({ x, y });
+    } else {
+      cancel();
+    }
+  });
+
   const dispatch = useAppDispatch();
   const activeLayer = useAppSelector(selectActiveLayer);
 
@@ -26,23 +62,27 @@ const ColorPicker = () => {
   const brushColor = useAppSelector(selectBrushColor);
 
   const { open, close, isSupported } = useEyeDropper();
-  const color = activeLayer === "mask" ? maskColor : brushColor;
+  const color = activeLayer?.includes("mask") ? maskColor : brushColor;
 
   const handleColorChange = (color: string) => {
-    if (activeLayer === "mask") dispatch(setMaskColor(color));
-    if (activeLayer === "sketch" || activeLayer.startsWith("controlnet"))
-      dispatch(setBrushColor(color));
+    if (onColorChange) onColorChange(color);
+    else {
+      if (activeLayer?.includes("mask")) dispatch(setMaskColor(color));
+      if (activeLayer === "sketch" || activeLayer.startsWith("controlnet"))
+        dispatch(setBrushColor(color));
+    }
   };
-  const handleKeydown: KeyboardEventHandler<HTMLElement> = useCallback(
+  const handleKeydown: KeyboardEventHandler = useCallback(
     (e) => {
       if (
         e.key === "Escape" ||
         (e.key.toLowerCase() === "p" && !e.shiftKey && !e.altKey && !e.ctrlKey)
       ) {
-        dispatch(toggleColorPickerVisibility());
+        if (onClose) onClose();
+        else dispatch(toggleColorPickerVisibility());
       }
     },
-    [dispatch]
+    [dispatch, onClose]
   );
 
   useGlobalKeydown({ handleKeydown });
@@ -56,23 +96,38 @@ const ColorPicker = () => {
       close();
     }
   };
+
   return (
-    <div className="flex flex-col rounded p-4 bg-black/90 backdrop-blur-md border-neutral-900/60 border">
-      <HexAlphaColorPicker
-        color={color}
-        onChange={handleColorChange}
-        onKeyDown={handleKeydown}
-      />
-      {isSupported() && (
-        <button
-          className="border-neutral-700 bg-neutral-900 flex justify-center rounded rounded-t-none"
-          onClick={pickColor}
+    <animated.div
+      ref={containerRef}
+      className="block group absolute z-10 shadow-lg shadow-black/50 border border-neutral-800/40 rounded-md bg-black/60 backdrop-blur-md p-7 cursor-move"
+      // style={{ left: posX, top: posY }}
+      style={style}
+      {...bind()}
+    >
+      <div
+        onClick={() => onClose?.()}
+        className="opacity-0 pointer-events-none group-hover:opacity-100 hover:opacity-100 group-hover:pointer-events-auto absolute top-[-19px] end-[-19px] cursor-pointer select-none rounded-full bg-black/80 hover:bg-neutral-800/70 backdrop-blur-md size-10 transition-all duration-300 flex justify-center items-center"
+      >
+        <Cross2Icon />
+      </div>
+      <div className="flex flex-col">
+        <HexAlphaColorPicker
+          color={color}
+          onChange={handleColorChange}
           onKeyDown={handleKeydown}
-        >
-          <BiSolidEyedropper />
-        </button>
-      )}
-    </div>
+        />
+        {isSupported() && (
+          <button
+            className="border-neutral-700/60 bg-neutral-900/60 flex justify-center rounded-md rounded-t-none"
+            onClick={pickColor}
+            onKeyDown={handleKeydown}
+          >
+            <BiSolidEyedropper />
+          </button>
+        )}
+      </div>
+    </animated.div>
   );
 };