+++ assets/js/src/components/Checkbox.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef, useId } from "react";
+import { forwardRef, useId } from "react";
 import * as RadixCheckbox from "@radix-ui/react-checkbox";
 import { CheckIcon } from "@radix-ui/react-icons";
 import Label from "./Label";
@@ -10,22 +10,26 @@ interface Props {
 
 const Checkbox = forwardRef<
   React.ElementRef<typeof RadixCheckbox.Root>,
-  Props & Omit<RadixCheckbox.CheckboxProps, "id">
->(({ children, onChange, ...rest }, _ref) => {
+  Props & Omit<RadixCheckbox.CheckboxProps, "id" | "value">
+>(({ children, onChange, value, ...rest }, _ref) => {
   const id = useId();
 
   return (
-    <div className="flex items-start md:items-center gap-4 justify-between flex-col md:flex-row">
+    <div className="group flex items-start md:items-center gap-4 justify-between flex-col md:flex-row">
       {children && (
-        <Label className="Label select-none" htmlFor={id}>
+        <Label
+          className="Label select-none cursor-pointer group-hover:text-neutral-200 transition-colors"
+          htmlFor={id}
+        >
           {children}
         </Label>
       )}
       <RadixCheckbox.Root
-        className="w-8 h-8 shrink-0 flex items-center  justify-center bg-neutral-900 p-0 rounded bg-neutral-800/80 hover:bg-neutral-800 border border-neutral-700"
+        className="w-6 h-6 shrink-0 flex items-center  justify-center bg-neutral-900 p-0 rounded bg-neutral-800/80 hover:bg-neutral-800 border border-neutral-700"
         onCheckedChange={onChange}
         id={id}
-        checked={rest.value}
+        value={value as RadixCheckbox.CheckboxProps["value"]}
+        checked={value}
         {...rest}
       >
         <RadixCheckbox.Indicator className="text-primary">