+++ test/ex_sd/sd/comfy_prompt_test.exs
@@ -175,7 +175,7 @@ defmodule ExSd.Sd.ComfyPromptTest do
       %ExSd.Sd.ControlNetArgs{
         model: "test_cn_model1",
         module: "none",
-        input_image: "data:image/png;base64,input_image"
+        image: "data:image/png;base64,input_image"
       },
       %ExSd.Sd.ControlNetArgs{
         model: "test_cn_model2",